Изменение автоматизированной сборки и живого сервера
