+++ form.html
@@ -320,6 +320,6 @@
         <button class="btn  btn--submit" type="button">Закрыть окно</button>
       </div>
     </section>
-    <script src="js/scripts.js"></script>
+    <script src="js/scripts.min.js"></script>
   </body>
 </html>
