+++ js/adaptive-menu.js
@@ -1,8 +1,10 @@
 ;(function() {
   'use strict';
 
-  var navToggler = document.getElementById('navToggler'),
-    mainNav = document.querySelector('.main-nav'),
+  var navToggler = document.getElementById('navToggler');
+  if (!navToggler) { return };
+
+  var mainNav = document.querySelector('.main-nav'),
     mainNavList = document.getElementById('mainNav'),
     pageHeader = document.querySelector('.page-header');
