+++ package.json
@@ -19,6 +19,7 @@
     "css-mqpacker": "^5.0.1",
     "del": "^2.2.2",
     "gulp": "3.9.1",
+    "gulp-concat": "^2.6.0",
     "gulp-csso": "^2.0.0",
     "gulp-imagemin": "^3.0.3",
     "gulp-plumber": "1.1.0",
