+++ photo.html
@@ -528,6 +528,6 @@
         </div>
       </div>
     </footer>
-    <script src="js/scripts.js"></script>
+    <script src="js/scripts.min.js"></script>
   </body>
 </html>
